Move refresh token functionality to AuthService
