+++ IOSAccessAssessment/Services/AuthService.swift
@@ -75,6 +75,64 @@ class AuthService {
         keychainService.removeValue(for: .refreshExpirationDate)
     }
     
+    func refreshToken() {
+        guard let refreshToken = keychainService.getValue(for: .refreshToken) else {
+            print("No refresh token found.")
+            return
+        }
+        
+        sendRefreshTokenRequest(refreshToken: refreshToken) { [weak self] result in
+            switch result {
+            case .success(let authResponse):
+                self?.storeAuthData(authResponse: authResponse)
+            case .failure(let error):
+                print("Failed to refresh token: \(error)")
+            }
+        }
+    }
+    
+    private func sendRefreshTokenRequest(
+        refreshToken: String,
+        completion: @escaping (Result<AuthResponse, NetworkError>) -> Void
+    ) {
+        guard let url = URL(string: "https://tdei-gateway-stage.azurewebsites.net/api/v1/refresh-token") else {
+            completion(.failure(.invalidURL))
+            return
+        }
+
+        var request = URLRequest(url: url)
+        request.httpMethod = "POST"
+        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
+        
+        request.httpBody = refreshToken.data(using: .utf8)
+
+        print("request: ", request)
+        
+        URLSession.shared.dataTask(with: request) { data, response, error in
+            if let error = error {
+                completion(.failure(.serverError(message: error.localizedDescription)))
+                return
+            }
+
+            guard let httpResponse = response as? HTTPURLResponse, (200...299).contains(httpResponse.statusCode) else {
+                completion(.failure(.invalidResponse))
+                return
+            }
+
+            guard let data = data else {
+                completion(.failure(.noData))
+                return
+            }
+
+            do {
+                let authResponse = try JSONDecoder().decode(AuthResponse.self, from: data)
+                completion(.success(authResponse))
+            } catch {
+                completion(.failure(.decodingError))
+            }
+        }.resume()
+    }
+    
     private func createRequest(username: String, password: String) -> URLRequest? {
         guard let url = URL(string: Constants.serverUrl) else { return nil }
         
@@ -112,7 +170,7 @@ class AuthService {
     ) {
         do {
             let authResponse = try JSONDecoder().decode(AuthResponse.self, from: data)
-            KeychainService().storeAuthData(authResponse: authResponse)
+            storeAuthData(authResponse: authResponse)
             completion(.success(authResponse))
         } catch {
             completion(.failure(.decodingError))
@@ -134,6 +192,16 @@ class AuthService {
         } catch {
             completion(.failure(.decodingError))
         }
+    }
+    
+    private func storeAuthData(authResponse: AuthResponse) {
+        keychainService.setValue(authResponse.accessToken, for: .accessToken)
+        let expirationDate = Date().addingTimeInterval(TimeInterval(authResponse.expiresIn))
+        keychainService.setDate(expirationDate, for: .expirationDate)
+        
+        keychainService.setValue(authResponse.refreshToken, for: .refreshToken)
+        let refreshExpirationDate = Date().addingTimeInterval(TimeInterval(authResponse.refreshExpiresIn))
+        keychainService.setDate(refreshExpirationDate, for: .refreshExpirationDate)
     }
     
 }
